Rename ToolbarPlugin.tsx → ToolbarExtension.tsx
The file contains an extension, not a plugin. Use consistent naming
with AIExtension.ts and the other extension files.
