+++ examples/agent-example/src/plugins/ToolbarExtension.tsx
